split scopes into files
